requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
captcha==0.7.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==45.0.6
cycler==0.12.1
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
extra-streamlit-components==0.1.80
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-MySQL==1.6.0
flask-mysql-connector==1.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.4
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.74.0
grpcio-status==1.71.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lxml==6.0.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.6.3
mysql-connector-python==9.2.0
narwhals==2.0.1
numpy==2.2.2
oauth2client==4.1.3
oauthlib==3.3.1
openai==0.28.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.9.0
PyMuPDF==1.26.3
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
streamlit==1.47.1
streamlit-authenticator==0.4.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.0
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
yarl==1.20.1
